Changed tests to account for new behavior of when slack invites are sent
Instead of running upon the account being registered, the slack invite is sent when the email is verified. The existing test has been changed to anticipate the slack invite no longer being sent on on registration and a new test has been added to check for the test when the email has been verified.